Risks packages/EU AI Act Compliance/data governance/pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data_governance"
version = "0.1.0"
description = "A data governance project for High-Risk AI (EU AI Act compliant)"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "great_expectations",
    "fairlearn",
    "scikit-learn",
    "presidio-analyzer",
    "presidio-anonymizer",
    "openlineage-python",
    "mcp",
    "spacy",
    "codecarbon",
    "pydantic",
    "duckdb",
    "reportlab"
]

[project.urls]
"Homepage" = "https://github.com/username/data_governance"

[project.scripts]
data-gov = "data_governance.main:main"
data-gov-server = "data_governance.server:main"

[tool.setuptools.packages.find]
where = ["src"]
